requirements.txt
asttokens==3.0.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
decorator==5.2.1
executing==2.2.1
fonttools==4.60.1
idna==3.11
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonpickle==4.1.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
narwhals==2.10.2
networkx==3.5
numpy==2.3.4
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
plotly==6.4.0
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.23
PyGithub==2.8.1
Pygments==2.19.2
PyJWT==2.10.1
PyNaCl==1.6.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-louvain==0.16
pytz==2025.2
pyvis==0.3.2
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
stack-data==0.6.3
threadpoolctl==3.6.0
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
